refactor(socceranalyzer): change project structure
