+++ socceranalyzer/RobotSSL/robot_ssl.py
@@ -1,0 +1,3 @@
+class RobotSSL:
+    def __init__(self) -> None:
+        pass
