+++ socceranalyzer/RobotVSS/robot_vss.py
@@ -1,0 +1,3 @@
+class RobotVSS:
+    def __init__(self) -> None:
+        pass
