+++ socceranalyzer/agent2D/player/agent.py
@@ -1,4 +1,6 @@
-class Agent:
+from socceranalyzer.common.entity.abstract_player import AbstractPlayer
+
+class Agent2D(AbstractPlayer):
     def __init__(self, team_name : str = None, team_side: str = None, number_id: int = None):
         """
             A class to represent a agent that is playing the match
