+++ socceranalyzer/common/entity/abstract_player.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 from socceranalyzer.common.geometric.point import Point
 
-class AbstractEntity(metaclass=ABCMeta):
+class AbstractPlayer(metaclass=ABCMeta):
     @abstractmethod
     def positionAt(self, cycle : int) -> Point:
         pass
