--- socceranalyzer/common/entity/abstract_robot.py
@@ -1,3 +1,0 @@
-class AbstractRobot:
-    def __init__(self):
-        pass
